# Find-DbaDatabase, performance