<!-- Below information IS REQUIRED with every PR -->
## Type of Change
<!-- What type of change does your code introduce -->
 - [ ] Bug fix (non-breaking change, fixes #<!--issue number--> )
 - [ ] New feature (non-breaking change, adds functionality, fixes #<!--issue number--> )
 - [ ] Breaking change (effects multiple commands or functionality, fixes #<!--issue number--> )
 - [x] Ran manual Pester test and has passed (`.\tests\manual.pester.ps1)
 - [ ] Adding code coverage to existing functionality
 - [ ] Pester test is included
 - [ ] If new file reference added for test, has is been added to github.com/sqlcollaborative/appveyor-lab ?
 - [ ] Nunit test is included
 - [ ] Documentation
 - [ ] Build system
 
### Purpose
I can only sum it up with "speeeeeeeed" ;-)
I know we favor smo methods if are there but counting tables, views and sproc via smo is too damn slow.

### Approach
Plain old query on the db, those DMVs should never change

### Screenshots
```
  Describing Find-DbaDatabase Unit Tests

    Context Validate parameters
      [+] Should only contain our specific parameters 171ms

  Describing Find-DbaDatabase Integration Tests

    Context Command actually works
      [+] Should return correct properties 36.57s
      [+] Should return true if Database Master is Found 33.41s
      [+] Should return true if Creation Date of Master is '4/8/2003 9:13:36 AM' 33ms
      [+] Should return true if Executed Against 2 instances: WIN-V28FT1UUD3E and WIN-V28FT1UUD3E\base2017 63.84s
      [+] Should return true if Database Found via Property Filter 48.77s
Tests completed in 182.78s
Tests Passed: 6, Failed: 0, Skipped: 0, Pending: 0, Inconclusive: 0
```
VS

```
  Describing Find-DbaDatabase Unit Tests

    Context Validate parameters
      [+] Should only contain our specific parameters 167ms

  Describing Find-DbaDatabase Integration Tests

    Context Command actually works
      [+] Should return correct properties 474ms
      [+] Should return true if Database Master is Found 109ms
      [+] Should return true if Creation Date of Master is '4/8/2003 9:13:36 AM' 21ms
      [+] Should return true if Executed Against 2 instances: WIN-V28FT1UUD3E and WIN-V28FT1UUD3E\base2017 221ms
      [+] Should return true if Database Found via Property Filter 320ms
Tests completed in 1.31s
Tests Passed: 6, Failed: 0, Skipped: 0, Pending: 0, Inconclusive: 0
```