@wsmelton 2020-08-11
We can speed it up even more by just removing those properties altogether. Why would I need to know how many tables are in a database if I'm just trying to find the database itself?